fix right stories
